RESTWS-398 - Moving create/update changes to Patient/Person resource from 1.9 to 1.8
