+++ omod-1.8/src/main/java/org/openmrs/module/webservices/rest/web/v1_0/resource/openmrs1_8/PersonResource1_8.java
@@ -44,7 +44,7 @@ import org.openmrs.util.OpenmrsUtil;
 /**
  * {@link Resource} for Person, supporting standard CRUD operations
  */
-@Resource(name = RestConstants.VERSION_1 + "/person", order = 1, supportedClass = Person.class, supportedOpenmrsVersions = {"1.8.*"})
+@Resource(name = RestConstants.VERSION_1 + "/person", order = 1, supportedClass = Person.class, supportedOpenmrsVersions = {"1.8.*", "1.9.*"})
 //order must be greater than that for PatientResource(order=0) RESTWS-273
 public class PersonResource1_8 extends DataDelegatingCrudResource<Person> {
 	
@@ -117,19 +117,28 @@ public class PersonResource1_8 extends DataDelegatingCrudResource<Person> {
 		description.addProperty("attributes");
 		return description;
 	}
-	
-	/**
-	 * @throws ResourceDoesNotSupportOperationException
-	 * @see org.openmrs.module.webservices.rest.web.resource.impl.BaseDelegatingResource#getUpdatableProperties()
-	 */
-	@Override
-	public DelegatingResourceDescription getUpdatableProperties() throws ResourceDoesNotSupportOperationException {
-		DelegatingResourceDescription description = super.getUpdatableProperties();
-		description.removeProperty("age");
-		description.addProperty("preferredName");
-		description.addProperty("preferredAddress");
-		return description;
-	}
+
+    /**
+     * @throws org.openmrs.module.webservices.rest.web.response.ResourceDoesNotSupportOperationException
+     *
+     * @see org.openmrs.module.webservices.rest.web.resource.impl.BaseDelegatingResource#getUpdatableProperties()
+     */
+    @Override
+    public DelegatingResourceDescription getUpdatableProperties() throws ResourceDoesNotSupportOperationException {
+        DelegatingResourceDescription description = new DelegatingResourceDescription();
+        description.addProperty("age");
+        description.addProperty("gender");
+        description.addProperty("birthdate");
+        description.addProperty("birthdateEstimated");
+        description.addProperty("preferredName");
+        description.addProperty("preferredAddress");
+        description.addProperty("addresses");
+        description.addProperty("attributes");
+        description.addRequiredProperty("names");
+        description.addRequiredProperty("causeOfDeath");
+        description.addRequiredProperty("dead");
+        return description;
+    }
 	
 	/**
 	 * @see org.openmrs.module.webservices.rest.web.resource.impl.BaseDelegatingResource#getPropertiesToExposeAsSubResources()
@@ -149,50 +158,55 @@ public class PersonResource1_8 extends DataDelegatingCrudResource<Person> {
 	public static Set<PersonName> getNames(Person instance) {
 		return RestUtil.removeVoidedData(instance.getNames());
 	}
+
+    /**
+     * Sets names and marks the first one as preferred if none is marked. It also makes sure that
+     * only one name is marked as preferred and changes the rest to non-preferred.
+     * <p/>
+     * It takes the list so that the order is preserved.
+     *
+     * @param instance
+     * @param names
+     */
+    @PropertySetter("names")
+    public static void setNames(Person instance, List<PersonName> names) {
+        for (PersonName existingName : instance.getNames()) {
+            existingName.setPreferred(false);
+        }
+        setFirstNameAsPreferred(names);
+        for (PersonName name : names) {
+            PersonName existingName = getMatchingName(name, instance.getNames());
+            if (existingName != null) {
+                copyNameFields(existingName, name);
+            } else {
+                instance.addName(name);
+            }
+        }
+    }
 	
-	/**
-	 * Sets names and marks the first one as preferred if none is marked. It also makes sure that
-	 * only one name is marked as preferred and changes the rest to non-preferred.
-	 * <p>
-	 * It takes the list so that the order is preserved.
-	 * 
-	 * @param instance
-	 * @param names
-	 */
-	@PropertySetter("names")
-	public static void setNames(Person instance, List<PersonName> names) {
-		boolean hasPreferred = false;
-		for (PersonName name : names) {
-			if (name.isPreferred()) {
-				if (!hasPreferred) {
-					hasPreferred = true;
-				} else {
-					name.setPreferred(false);
-				}
-			}
-		}
-		
-		if (!hasPreferred) {
-			names.iterator().next().setPreferred(true);
-		}
-		
-		//Hibernate expects java.util.SortedSet
-		instance.setNames(new TreeSet<PersonName>(names));
-	}
-	
-	/**
-	 * Sets attributes on the given person.
-	 * 
-	 * @param instance
-	 * @param names
-	 */
-	@PropertySetter("attributes")
-	public static void setAttributes(Person instance, List<PersonAttribute> attrs) {
-		for (PersonAttribute attr : attrs)
-			instance.addAttribute(attr);
-	}
-	
-	/**
+    /**
+     * Sets attributes on the given person.
+     *
+     * @param instance
+     * @param attrs
+     */
+    @PropertySetter("attributes")
+    public static void setAttributes(Person instance, List<PersonAttribute> attrs) {
+        for (PersonAttribute attr : attrs) {
+            PersonAttribute existingAttribute = instance.getAttribute(Context.getPersonService().getPersonAttributeTypeByUuid(attr.getAttributeType().getUuid()));
+            if (existingAttribute != null) {
+                if (attr.getValue() == null) {
+                    instance.removeAttribute(existingAttribute);
+                } else {
+                    existingAttribute.setValue(attr.getValue());
+                }
+            } else {
+                instance.addAttribute(attr);
+            }
+        }
+    }
+
+    /**
 	 * Returns non-voided addresses for a person
 	 * 
 	 * @param instance
@@ -202,37 +216,31 @@ public class PersonResource1_8 extends DataDelegatingCrudResource<Person> {
 	public static Set<PersonAddress> getAddresses(Person instance) {
 		return RestUtil.removeVoidedData(instance.getAddresses());
 	}
-	
-	/**
-	 * Sets addresses and marks the first one as preferred if none is marked. It also makes sure
-	 * that only one address is marked as preferred and changes the rest to non-preferred.
-	 * <p>
-	 * It takes the list so that the order is preserved.
-	 * 
-	 * @param instance
-	 * @param addresses
-	 */
-	@PropertySetter("addresses")
-	public static void setAddresses(Person instance, List<PersonAddress> addresses) {
-		boolean hasPreferred = false;
-		for (PersonAddress address : addresses) {
-			if (address.isPreferred()) {
-				if (!hasPreferred) {
-					hasPreferred = true;
-				} else {
-					address.setPreferred(false);
-				}
-			}
-		}
-		
-		if (!hasPreferred) {
-			addresses.iterator().next().setPreferred(true);
-		}
-		
-		//Hibernate expects java.util.SortedSet
-		instance.setAddresses(new TreeSet<PersonAddress>(addresses));
-	}
-	
+
+    /**
+     * Sets addresses and marks the first one as preferred if none is marked. It also makes sure
+     * that only one address is marked as preferred and changes the rest to non-preferred.
+     * <p/>
+     * It takes the list so that the order is preserved.
+     *
+     * @param instance
+     * @param addresses
+     */
+    @PropertySetter("addresses")
+    public static void setAddresses(Person instance, List<PersonAddress> addresses) {
+        for (PersonAddress existingAddress : instance.getAddresses()) {
+            existingAddress.setPreferred(false);
+        }
+        setFirstAddressAsPreferred(addresses);
+        for (PersonAddress address : addresses) {
+            PersonAddress existingAddress = getMatchingAddress(address, instance.getAddresses());
+            if (existingAddress != null) {
+                copyAddressFields(existingAddress, address);
+            } else {
+                instance.addAddress(address);
+            }
+        }
+    }
 	/**
 	 * Sets the preferred name for a person. If no name exists new name is set as preferred.
 	 * 
@@ -355,4 +363,86 @@ public class PersonResource1_8 extends DataDelegatingCrudResource<Person> {
 		
 		return person.getPersonName().getFullName();
 	}
+
+    private static void copyNameFields(PersonName existingName, PersonName personName) {
+        existingName.setPreferred(personName.getPreferred());
+        existingName.setPrefix(personName.getPrefix());
+        existingName.setGivenName(personName.getGivenName());
+        existingName.setMiddleName(personName.getMiddleName());
+        existingName.setFamilyNamePrefix(personName.getFamilyNamePrefix());
+        existingName.setFamilyName(personName.getFamilyName());
+        existingName.setFamilyName2(personName.getFamilyName2());
+        existingName.setFamilyNameSuffix(personName.getFamilyNameSuffix());
+        existingName.setDegree(personName.getDegree());
+    }
+
+    private static void setFirstNameAsPreferred(List<PersonName> personNames) {
+        boolean hasPreferred = false;
+        for (PersonName name : personNames) {
+            if (name.isPreferred()) {
+                if (!hasPreferred) {
+                    hasPreferred = true;
+                } else {
+                    name.setPreferred(false);
+                }
+            }
+        }
+        if (!hasPreferred) {
+            personNames.iterator().next().setPreferred(true);
+        }
+    }
+
+    private static PersonName getMatchingName(PersonName personName, Set<PersonName> personNames) {
+        for (PersonName existingName : personNames) {
+            String uuid = personName.getUuid();
+            if (uuid != null && uuid.equals(existingName.getUuid())) {
+                return existingName;
+            }
+        }
+        return null;
+    }
+
+    private static PersonAddress getMatchingAddress(PersonAddress personAddress, Set<PersonAddress> personAddresses) {
+        for (PersonAddress existingAddress : personAddresses) {
+            if (personAddress.getUuid().equals(existingAddress.getUuid())) {
+                return existingAddress;
+            }
+        }
+        return null;
+    }
+
+    private static void copyAddressFields(PersonAddress existingAddress, PersonAddress address) {
+        existingAddress.setPreferred(address.getPreferred());
+        existingAddress.setAddress1(address.getAddress1());
+        existingAddress.setAddress2(address.getAddress2());
+        existingAddress.setAddress3(address.getAddress3());
+        existingAddress.setAddress4(address.getAddress4());
+        existingAddress.setAddress5(address.getAddress5());
+        existingAddress.setAddress6(address.getAddress6());
+        existingAddress.setCityVillage(address.getCityVillage());
+        existingAddress.setCountry(address.getCountry());
+        existingAddress.setStateProvince(address.getStateProvince());
+        existingAddress.setCountyDistrict(address.getCountyDistrict());
+        existingAddress.setPostalCode(address.getPostalCode());
+        existingAddress.setLatitude(address.getLatitude());
+        existingAddress.setLongitude(address.getLongitude());
+        existingAddress.setDateCreated(address.getDateCreated());
+    }
+
+    private static void setFirstAddressAsPreferred(List<PersonAddress> addresses) {
+        boolean hasPreferred = false;
+        for (PersonAddress address : addresses) {
+            if (address.isPreferred()) {
+                if (!hasPreferred) {
+                    hasPreferred = true;
+                } else {
+                    address.setPreferred(false);
+                }
+            }
+        }
+        if (!hasPreferred) {
+            addresses.iterator().next().setPreferred(true);
+        }
+    }
+
 }
