+++ omod-1.8/src/test/resources/update_patient.json
@@ -33,7 +33,8 @@
                 "address6": "address 6",
                 "latitude": null,
                 "longitude": null,
-                "voided": false
+                "voided": false,
+                "dateCreated": "2013-12-08T00:00:00.000+0530"
             }
         ],
         "attributes": [
