--- omod-1.9/src/main/java/org/openmrs/module/webservices/rest/web/v1_0/resource/openmrs1_9/PatientResource1_9.java
@@ -1,68 +1,0 @@
-/**
- * The contents of this file are subject to the OpenMRS Public License
- * Version 1.0 (the "License"); you may not use this file except in
- * compliance with the License. You may obtain a copy of the License at
- * http://license.openmrs.org
- *
- * Software distributed under the License is distributed on an "AS IS"
- * basis, WITHOUT WARRANTY OF ANY KIND, either express or implied. See the
- * License for the specific language governing rights and limitations
- * under the License.
- *
- * Copyright (C) OpenMRS, LLC.  All Rights Reserved.
- */
-package org.openmrs.module.webservices.rest.web.v1_0.resource.openmrs1_9;
-
-import org.openmrs.Patient;
-import org.openmrs.Person;
-import org.openmrs.api.context.Context;
-import org.openmrs.module.webservices.rest.SimpleObject;
-import org.openmrs.module.webservices.rest.web.ConversionUtil;
-import org.openmrs.module.webservices.rest.web.RequestContext;
-import org.openmrs.module.webservices.rest.web.RestConstants;
-import org.openmrs.module.webservices.rest.web.annotation.Resource;
-import org.openmrs.module.webservices.rest.web.api.RestService;
-import org.openmrs.module.webservices.rest.web.representation.Representation;
-import org.openmrs.module.webservices.rest.web.resource.impl.DelegatingResourceDescription;
-import org.openmrs.module.webservices.rest.web.response.ResponseException;
-import org.openmrs.module.webservices.rest.web.v1_0.resource.openmrs1_8.PatientResource1_8;
-
-import java.util.*;
-
-/**
- * {@link org.openmrs.module.webservices.rest.web.annotation.Resource} for Patients, supporting standard CRUD operations
- */
-@Resource(name = RestConstants.VERSION_1 + "/patient", supportedClass = Patient.class, supportedOpenmrsVersions = {"1.9.*"})
-public class PatientResource1_9 extends PatientResource1_8 {
-	
-	public PatientResource1_9() {
-	}
-	
-	/**
-	 * @see org.openmrs.module.webservices.rest.web.resource.impl.BaseDelegatingResource#getUpdatableProperties()
-	 */
-	@Override
-	public DelegatingResourceDescription getUpdatableProperties() {
-		DelegatingResourceDescription description = new DelegatingResourceDescription();
-        description.addRequiredProperty("person");
-		return description;
-	}
-	
-    @Override
-    public Object update(String uuid, SimpleObject propertiesToUpdate, RequestContext context) throws ResponseException {
-        if(propertiesToUpdate.get("person") == null) {
-            return super.update(uuid, propertiesToUpdate, context);
-        }
-        Patient patient = getPatientForUpdate(uuid, propertiesToUpdate);
-        patient = save(patient);
-        return ConversionUtil.convertToRepresentation(patient, Representation.DEFAULT);
-    }
-
-    public Patient getPatientForUpdate(String uuid, Map<String, Object> propertiesToUpdate) {
-        Patient patient = getByUniqueId(uuid);
-        PersonResource1_9 personResource = (PersonResource1_9) Context.getService(RestService.class).getResourceBySupportedClass(Person.class);
-        personResource.setConvertedProperties(patient, (Map<String, Object>) propertiesToUpdate.get("person"), personResource.getUpdatableProperties(), false);
-        return patient;
-    }
-
-}
